Cleaning Program, Ref Assignments, Repo cleanup
Created a folder for the pre-processing program and requirements for the program and moved cleaning notebook into that folder.

Also created a folder for reference assignments, so they are available if needed for the project.

Also cleaned the repo a little, and created a gitignore to make git ignore the dataset.